+++ homeassistant/components/simplisafe/manifest.json
@@ -4,7 +4,7 @@
   "config_flow": true,
   "documentation": "https://www.home-assistant.io/integrations/simplisafe",
   "requirements": [
-    "simplisafe-python==5.0.1"
+    "simplisafe-python==5.1.0"
   ],
   "dependencies": [],
   "codeowners": [
