+++ requirements_all.txt
@@ -1769,7 +1769,7 @@ shodan==1.19.0
 simplepush==1.1.4
 
 # homeassistant.components.simplisafe
-simplisafe-python==5.0.1
+simplisafe-python==5.1.0
 
 # homeassistant.components.sisyphus
 sisyphus-control==2.2.1
