+++ requirements_test_all.txt
@@ -546,7 +546,7 @@ rxv==0.6.0
 samsungctl[websocket]==0.7.1
 
 # homeassistant.components.simplisafe
-simplisafe-python==5.0.1
+simplisafe-python==5.1.0
 
 # homeassistant.components.sleepiq
 sleepyq==0.7
